import/requirements.txt
python-qbittorrent
flask
werkzeug
wtforms